backend/requirements.txt
fastapi
uvicorn
openai
pyodbc
azure-search-documents
azure-core
azure-identity
azure-keyvault-secrets
